requirements.txt
absl-py
astor
bleach
bottle
click
certifi
chardet
flask
future
gast
grpcio
gunicorn
h5py
html5lib
itsdangerous
idna
Jinja2
Keras
line-bot-sdk
Markdown
MarkupSafe
numpy<2.0.0
oauthlib
pillow
protobuf
requests
six
tensorboard
tensorboard-data-server
tensorflow-cpu==2.15.0
termcolor
urllib3
Werkzeug
wheel